requirements.txt
streamlit==1.31.1
anthropic==0.18.1
python-dotenv==1.0.1
fastapi==0.110.0
pydantic==2.6.3
python-dotenv==1.0.1
pdfplumber==0.10.4
google-cloud-vision==3.7.1
python-docx==1.1.0